+++ src/extension-firefox/content.js
@@ -3,7 +3,7 @@
 const debug = false;
 
 // These constants are defined in '../constants.js'
-const EXTENSION_ELEMENT_NAME = 'skip-to-content';
+const EXTENSION_ELEMENT_NAME = 'skip-to-content-extension';
 const ATTR_SKIP_TO_DATA      = 'data-skipto';
 
 // Define browser specific APIs for Opera, Firefox and Chrome