+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Version history for SkipTo.js
 
 ## Version 5.7.4
-* Added support for dark mode in overlays
+* Added support for dark mode in highlight overlay
+* Improved programmatic constant management in source code
+* Updated some of the custom element names
 
 ## Version 5.7.3
 * Disabled support for dark mode in overlays