+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added support for dark mode in highlight overlay
 * Improved programmatic constant management in source code
 * Updated some of the custom element names
+* Updated FAQ and home page
 
 ## Version 5.7.3
 * Disabled support for dark mode in overlays